duplicate operations for new runner
